[tests] get rid of test/controllers
It was a mixed bag of functional and integration tests that had
class name clashes with other integration/functional tests.
