+++ test/functional/api_docs/proxy_controller_test.rb
