+++ test/functional/developer_portal/admin/messages/trash_controller_test.rb
@@ -1,10 +1,23 @@
 require 'test_helper'
 
 class DeveloperPortal::Admin::Messages::TrashControllerTest < DeveloperPortal::ActionController::TestCase
-  with_options :controller => 'messages/trash' do |test|
-    test.should route(:get, '/admin/messages/trash').to :action => 'index'
-    test.should route(:get, '/admin/messages/trash/42').to :action => 'show', :id => '42'
-    test.should route(:delete, '/admin/messages/trash/42').to :action => 'destroy', :id => '42'
-    test.should route(:delete, '/admin/messages/trash').to :action => 'empty'
+
+  def setup
+    provider = FactoryGirl.create(:provider_account)
+    @user    = FactoryGirl.create(:user, account: provider)
+
+    host! provider.domain
+
+    login_as @user
+  end
+
+  def test_index
+    get :index
+
+    assigned_drop_variables = assigns(:_assigned_drops).keys
+
+    assert :success
+    assert assigned_drop_variables.include?('messages')
+    assert assigned_drop_variables.include?('pagination')
   end
 end
