+++ test/functional/master/providers/switches_controller_test.rb
