+++ test/functional/provider/admin/api_docs_controller_test.rb
