+++ test/functional/provider/admin/dashboard/new_accounts_controller_test.rb
