+++ test/functional/provider/admin/dashboard/potential_upgrades_controller_test.rb
