+++ test/functional/provider/admin/dashboard/service/hits_controller_test.rb
