+++ test/functional/provider/admin/dashboard/service/top_traffic_controller_test.rb
