+++ test/functional/provider/admin/liquid_docs_controller_test.rb
