+++ test/functional/provider/admin/user/notification_preferences_controller_test.rb
