+++ test/integration/admin/cms/switches_controller_test.rb
